refactor: renamed B2C_account_top_up
