+++ tests/unit/b2c_account_top_up/test_b2c_account_top_up.py
@@ -9,7 +9,7 @@ from unittest.mock import MagicMock
 from mpesakit.auth import TokenManager
 from mpesakit.http_client import HttpClient
 
-from mpesakit.B2C_account_top_up import (
+from mpesakit.b2c_account_top_up import (
     B2CAccountTopUp,
     B2CAccountTopUpRequest,
     B2CAccountTopUpResponse,
